requirements.txt
apify-client 
dotenv
requests
openpyxl
openai
httpx
ruamel.yaml
python-docx
flet
nuitka